se suben cambios y desarrollos del iniciar procesamiento pedido

+++ TP_Servidor/src/entities/ItemFaltantePedidoEntity.java
@@ -49,7 +49,7 @@ public class ItemFaltantePedidoEntity implements Serializable {
 	private ColorEntity color;
 	
 	@Column(name="cantidadFaltante", nullable=false)
-	private int cantidadFaltante;
+	private float cantidadFaltante;
 	
 
 	public Integer getIdItemFaltantePedido() {
@@ -84,11 +84,11 @@ public class ItemFaltantePedidoEntity implements Serializable {
 		this.color = color;
 	}
 
-	public int getCantidadFaltante() {
+	public float getCantidadFaltante() {
 		return cantidadFaltante;
 	}
 
-	public void setCantidadFaltante(int cantidadFaltante) {
+	public void setCantidadFaltante(float cantidadFaltante) {
 		this.cantidadFaltante = cantidadFaltante;
 	}
